requirements.txt
Flask==1.0.2
fluent-logger
gunicorn==19.9.0
joblib==0.13.1
lime==0.1.1.33
numpy==1.15.4
pandas==0.23.4
scikit-learn==0.20.2
mlflow==0.8.2
